+++ src/journey.js
@@ -21,6 +21,13 @@
         // Variables: Configuration
         _configuration = {},
 
+        // Variables: Enums
+        _enum_KeyCodes = {
+            escape: 27,
+            left: 37,
+            right: 39
+        },
+
         // Variables: Strings
         _string = {
             empty: "",
@@ -279,6 +286,40 @@
         }
 
         return result;
+    }
+
+
+    /*
+     * ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------
+     * Document Events
+     * ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------
+     */
+
+    function buildDocumentEvents( addEvents ) {
+        addEvents = isDefined( addEvents ) ? addEvents : true;
+
+        var documentFunc = addEvents ? _parameter_Document.addEventListener : _parameter_Document.removeEventListener;
+
+        if ( _configuration.shortcutKeysEnabled ) {
+            documentFunc( "keydown", onWindowKeyDown );
+        }
+    }
+
+    function onWindowKeyDown( e ) {
+        if ( _this.isOpen() ) {
+            if ( e.keyCode === _enum_KeyCodes.escape ) {
+                e.preventDefault();
+                onDialogClose();
+
+            } else if ( e.keyCode === _enum_KeyCodes.left ) {
+                e.preventDefault();
+                onDialogPrevious();
+
+            } else if ( e.keyCode === _enum_KeyCodes.right ) {
+                e.preventDefault();
+                onDialogNext();
+            }
+        }
     }
 
 
@@ -561,6 +602,7 @@
         _configuration.nextButtonText = getDefaultString( _configuration.nextButtonText, "Next" );
         _configuration.finishButtonText = getDefaultString( _configuration.finishButtonText, "Finish" );
         _configuration.showCloseButton = getDefaultBoolean( _configuration.showCloseButton, true );
+        _configuration.shortcutKeysEnabled = getDefaultBoolean( _configuration.shortcutKeysEnabled, true );
     }
 
 
@@ -659,6 +701,7 @@
             renderDisabledBackground();
             renderDialog();
             getElements();
+            buildDocumentEvents();
         } );
 
         if ( !isDefined( _parameter_Window.$journey ) ) {