+++ docs/CHANGE_LOG.md
@@ -6,7 +6,8 @@
 - Added CSS ":root" variable "--dialog-close-button-border-color", which states the border color to use for the close button.
 - Added CSS ":root" variable "--dialog-close-button-hover-background-color", which states the background color to use for the close button when hovered.
 - Renamed one or two of the other ":root" variables for the Close button.
-- 
+- Added shortcut keys "Left", "Right" and "Escape", which move to the previous/next element in the journey, and close the dialog.
+- Added a new configuration option "shortcutKeysEnabled" (defaults to true), which states if the shortcut keys (for navigation) are enabled.
 
 <br>